+++ config_manager/config_defaults.cpp
@@ -31,6 +31,8 @@ namespace cfg {
 	char robonomics_connectivity_host[LEN_ROBONOMICS_CONNECTIVITY_HOST] = "";
 	char robonomics_connectivity_hosts[LEN_ROBONOMICS_CONNECTIVITY_HOSTS] = "";
 	bool rws_auto_register = true;
+	unsigned rws_group_mode = RWS_GROUP_STANDALONE;
+	char rws_group_id[LEN_RWS_GROUP_ID] = "";
 	char rws_devices_extra[LEN_RWS_DEVICES_EXTRA] = "";
 	char rws_devices_registered_hash[LEN_RWS_DEVICES_REGISTERED_HASH] = "";
 	char private_key[LEN_PRIVATE_KEY] = "Not Set";